third_party/WebKit/LayoutTests/fast/forms/week-multiple-fields/week-multiple-fields-ax-aria-attributes.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<input id="test" type="week" value="2012-W10">
<script>
description('This test checks aria-help attribute of fields in multiple fields week input UI.');

function focusedFieldValueDescription()
{
    var element = accessibilityController.focusedElement;
    return element.description + ', ' +  element.valueDescription + ', ' + element.minValue + ', ' + element.maxValue;
}

var testInput = document.getElementById('test');

if (!window.accessibilityController || !window.eventSender)
    debug('Please run inside DRT or WTR.');
else {
    debug('Non-empty value');
    testInput.focus();
    shouldBeEqualToString('focusedFieldValueDescription()', 'Week, AXValueDescription: 10, 1, 53');
    eventSender.keyDown('\t');
    shouldBeEqualToString('focusedFieldValueDescription()', 'Year, AXValueDescription: 2012, 1, 275760');

    debug('Empty value');
    eventSender.keyDown('Backspace');
    shouldBeEqualToString('focusedFieldValueDescription()', 'Year, AXValueDescription: ----, 1, 275760');
    eventSender.keyDown('\t', ['shiftKey']);
    eventSender.keyDown('Backspace');
    shouldBeEqualToString('focusedFieldValueDescription()', 'Week, AXValueDescription: --, 1, 53');

    debug('');
    testInput.remove();
}
</script>
</body>
</html>
