third_party/WebKit/LayoutTests/fast/forms/week-multiple-fields/week-multiple-fields-ax-aria-attributes-expected.txt

This test checks aria-help attribute of fields in multiple fields week input UI.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Non-empty value
PASS focusedFieldValueDescription() is "Week, AXValueDescription: 10, 1, 53"
PASS focusedFieldValueDescription() is "Year, AXValueDescription: 2012, 1, 275760"
Empty value
PASS focusedFieldValueDescription() is "Year, AXValueDescription: ----, 1, 275760"
PASS focusedFieldValueDescription() is "Week, AXValueDescription: --, 1, 53"

PASS successfullyParsed is true

TEST COMPLETE
